pyproject.toml
# plotpy setup configuration file

[build-system]
requires = ["setuptools", "wheel", "cython ~= 0.29.0", "numpy >= 1.17"]
build-backend = "setuptools.build_meta"

[project]
name = "PlotPy"
authors = [{ name = "Codra", email = "p.raybaut@codra.fr" }]
description = "Curve and image plotting tools for Python/Qt applications"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Human Machine Interfaces",
    "Topic :: Software Development :: User Interfaces",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: OS Independent",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8, <4"
dependencies = [
    "guidata>=3.1",
    "PythonQwt>=0.10",
    "NumPy>=1.17",
    "SciPy>=1.3",
    "Pillow",
    "tifffile",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/PlotPyStack/plotpy/"
Documentation = "https://plotpy.readthedocs.io/en/latest/"

[project.gui-scripts]
plotpy-tests = "plotpy.tests:run"

[project.optional-dependencies]
dev = ["black", "isort", "pylint", "Coverage", "Cython"]
doc = [
    "PyQt5",
    "sphinx",
    "myst_parser",
    "sphinx-copybutton",
    "sphinx_qt_documentation",
    "python-docs-theme",
]
test = ["pytest", "pytest-cov", "pytest-qt", "pytest-xvfb"]

[tool.setuptools.packages.find]
include = ["plotpy*"]

[tool.setuptools.package-data]
"*" = ["*.png", "*.svg", "*.mo", "*.cfg", "*.toml", "*.dcm", "*.ui", "*.json"]

[tool.setuptools.dynamic]
version = { attr = "plotpy.__version__" }
